Refactor fetchTasks API handler
* Convert it to a pure function
* Move it to a separate todos API file
